// types.ts
export enum ItemStatus {
  Pending = 'PENDING',
  Bought = 'BOUGHT',
}

export enum Unit {
  Kg = 'کیلوگرم',
  Gram = 'گرم',
  Liter = 'لیتر',
  Ml = 'میلی‌لیتر',
  Piece = 'عدد',
  Pack = 'بسته',
  Box = 'جعبه',
  Bottle = 'بطری',
  Can = 'قوطی',
  Jar = 'شیشه',
  Dozen = 'دوجین',
  Sheet = 'برگ',
  Roll = 'رول',
  Bag = 'کیسه',
  Bunch = 'دسته',
  Carton = 'کارتن',
  Sachet = 'ساشه',
  Bar = 'قالب',
}

export enum CafeCategory {
    Dairy = 'لبنیات',
    Produce = 'میوه و سبزیجات',
    Bakery = 'نان و شیرینی',
    DryGoods = 'خشکبار و حبوبات',
    Meat = 'گوشت و مرغ',
    Beverages = 'نوشیدنی‌ها',
    Cleaning = 'لوازم بهداشتی و نظافتی',
    Other = 'متفرقه',
}

export enum PaymentStatus {
  Paid = 'پرداخت شده',
  Due = 'پرداخت نشده',
}

export enum PaymentMethod {
  Cash = 'نقد',
  Card = 'کارت',
  Transfer = 'انتقال بانکی',
}
// ...
export interface AggregatedShoppingItem {
  name: string;
  unit: Unit;
  category: string;
  vendorId?: string;
  paymentMethod?: PaymentMethod;
  paymentStatus?: PaymentStatus;
  totalAmount: number;
  totalPrice: number;
  purchaseHistory: { date: Date, pricePerUnit: number }[];
}